refactor: modularize manage posts page
